+++ src/main/java/org/realityforge/replicant/server/ReplicantClient.java
@@ -1,0 +1,8 @@
+package org.realityforge.replicant.server;
+
+import java.util.List;
+
+public interface ReplicantClient
+{
+  void addChangeSet( List<EntityMessage> changeSet );
+}